Refactoring, UI responsiveness, fixed overlapping navbar menu bug
