+++ frontend/src/styles/TableToolBar.css
@@ -1,11 +1,9 @@
 .table-tool-bar {
   display: flex;
   align-items: center;
+  flex-wrap: wrap;
   flex: 1;
   justify-content: space-between;
-}
-
-.table-tool-bar .icon-buttons {
   padding: 0.6em;
 }
 
@@ -16,8 +14,8 @@
 
 .table-tool-bar .left-buttons {
   display: flex;
+  flex-wrap: wrap;
   align-items: center;
-  padding: 0;
   padding-left: 0.7em;
   margin: 0;
   gap: 0.4em;
